+++ src/lib/server/db/schema/courses/courses.ts
