+++ src/lib/server/db/schema/courses/lectureMaterials.ts
