+++ src/lib/server/db/schema/courses/lessons.ts
