+++ src/lib/server/db/schema/courses/quizzes.ts
